Seven Sages: fix the two cross-compiler breaks CI found
Neither is Enrich World, and neither reproduces on Clang.

SevenSagesRoomAoe.cpp - z64.h reaches <memory> transitively, and this file was
light enough on prior includes that it happened there first, inside extern "C".
That gives libstdc++'s parallel-algorithm templates C linkage, which GCC
rejects. Including <memory> ahead of the block makes the nested include a
no-op. libc++ has no such header in that path, hence macOS never saw it.

SevenSagesDekuShieldFire.cpp - FLAME_SCALE was a float passed to
func_8002836C's s16 scale parameter. MSVC reports the narrowing as C4244 and CI
builds with /WX. Now an s16 constant; 200 is exact either way.

Checked for more of both rather than waiting for CI to find them one at a time:
built a compile database and re-compiled every SevenSages and EnrichWorld file
with -Wfloat-conversion. FLAME_SCALE was the only hit of the class that breaks
MSVC. The integer-narrowing hits it also reported are Clang-only pedantry -
EnrichWorldStore.cpp carries three and compiled clean on Windows at [552/1705].

+++ soh/soh/Enhancements/SevenSages/SevenSagesDekuShieldFire.cpp
@@ -31,7 +31,11 @@ Color_RGBA8 sFlameEnvColor = { 255, 50, 0, 0 };
 
 // The stick scales its flame down as it burns out (`temp * 200.0f`). This one never burns out, so
 // it holds at the stick's full-strength value.
-constexpr float FLAME_SCALE = 200.0f;
+//
+// s16 rather than float because func_8002836C takes s16: as a float this narrowed implicitly at the
+// call site, which MSVC reports as C4244 and CI promotes to an error under /WX. 200 is exactly
+// representable either way, so the effect is unchanged.
+constexpr s16 FLAME_SCALE = 200;
 
 Player* CurrentPlayer() {
     if (!GameInteractor::IsSaveLoaded(true) || gPlayState == nullptr) {
